gpu works

+++ gpu/dws-gpu.h
@@ -1,2 +1,2 @@
-void dws_conv(double*, double*, double*, double*, int, int, int, int, int, int, int, int, int, int, int, int, double*);
+void dws_conv(float*, float*, float*, float*, int, int, int, int, int, int, int, int, int, int, int, int, float*);
 void init_conv(int, int, int, int, int, int, int, int, int, int, int, int);
